Removes redux and moves category functions to own component
