+++ src/components/CreateArticle.jsx
@@ -1,9 +1,9 @@
-import React, { useEffect, useRef } from "react";
+import React, { useEffect } from "react";
+import { useForm } from "react-hook-form";
 import { Header, Container, Form, Button } from "semantic-ui-react";
 import "react-semantic-ui-datepickers/dist/react-semantic-ui-datepickers.css";
-import _ from "lodash";
 import { Article } from "../modules/apiHelper";
-import { useForm } from "react-hook-form";
+import CategoryList from "./CategoryList";
 
 const CreateArticle = () => {
   useEffect(() => {
@@ -17,30 +17,6 @@ const CreateArticle = () => {
   const { register, handleSubmit, setValue } = useForm();
   const onSubmit = (article) => {
     Article.create({ article });
-  };
-
-  const options = [
-    "World News",
-    "Science & Enviroment",
-    "Business",
-    "Tech",
-    "Entertainment",
-    "Culture & Health",
-    "Sports",
-  ];
-
-  useEffect(() => {
-    optionsLoop();
-  }, []);
-
-  let categoryOptions = [];
-  const optionsLoop = () => {
-    options.forEach((options) => {
-      categoryOptions.push({
-        text: options,
-        value: _.camelCase(options),
-      });
-    });
   };
 
   return (
@@ -62,7 +38,7 @@ const CreateArticle = () => {
           name="title"
           data-cy="title-input"
           placeholder="Title"
-          onChange={async (e, { name, value }) => {
+          onChange={(e, { name, value }) => {
             setValue(name, value);
           }}
         />
@@ -70,7 +46,7 @@ const CreateArticle = () => {
           data-cy="journalist-input"
           placeholder="Journalists"
           name="journalist"
-          onChange={async (e, { name, value }) => {
+          onChange={(e, { name, value }) => {
             setValue(name, value);
           }}
         />
@@ -78,27 +54,16 @@ const CreateArticle = () => {
           data-cy="lede-input"
           placeholder="Lede"
           name="lede"
-          onChange={async (e, { name, value }) => {
+          onChange={(e, { name, value }) => {
             setValue(name, value);
           }}
         />
-        <Form.Group>
-          <Form.Select
-            data-cy="category-input"
-            placeholder="Category"
-            options={categoryOptions}
-            name="category"
-            value={categoryOptions.value}
-            onChange={async (e, { name, value }) => {
-              setValue(name, value);
-            }}
-          />
-        </Form.Group>
+        <CategoryList />
         <Form.TextArea
           data-cy="body-input"
           placeholder="Body"
           name="body"
-          onChange={async (e, { name, value }) => {
+          onChange={(e, { name, value }) => {
             setValue(name, value);
           }}
         />
